+++ docs/src/pages/components/Alert/NativeAlert.mdx
@@ -4,17 +4,17 @@ platform: native
 githubPath: bpk-component-alert
 ---
 
-import Readme from '../../../components/Readme';
-import NativeComponentScreenshots from '../../../components/Screenshots/NativeComponentScreenshots';
+import Readme from 'components/Readme';
+import NativeComponentScreenshots from 'components/Screenshots/NativeComponentScreenshots';
 
-import readme from '../../../../../backpack-react-native/lib/bpk-component-alert/README.md';
+import readme from 'backpack-react-native/lib/bpk-component-alert/README.md';
 
-import iosAlertDefault from '../../../../../backpack-react-native/lib/bpk-component-alert/screenshots/ios/default.png';
-import iosAlertCancelable from '../../../../../backpack-react-native/lib/bpk-component-alert/screenshots/ios/cancelable.png';
-import iosAlertThreeButton from '../../../../../backpack-react-native/lib/bpk-component-alert/screenshots/ios/three-button.png';
-import androidAlertDefault from '../../../../../backpack-react-native/lib/bpk-component-alert/screenshots/android/default.png';
-import androidAlertCancelable from '../../../../../backpack-react-native/lib/bpk-component-alert/screenshots/android/cancelable.png';
-import androidAlertThreeButton from '../../../../../backpack-react-native/lib/bpk-component-alert/screenshots/android/three-button.png';
+import iosAlertDefault from 'backpack-react-native/lib/bpk-component-alert/screenshots/ios/default.png';
+import iosAlertCancelable from 'backpack-react-native/lib/bpk-component-alert/screenshots/ios/cancelable.png';
+import iosAlertThreeButton from 'backpack-react-native/lib/bpk-component-alert/screenshots/ios/three-button.png';
+import androidAlertDefault from 'backpack-react-native/lib/bpk-component-alert/screenshots/android/default.png';
+import androidAlertCancelable from 'backpack-react-native/lib/bpk-component-alert/screenshots/android/cancelable.png';
+import androidAlertThreeButton from 'backpack-react-native/lib/bpk-component-alert/screenshots/android/three-button.png';
 
 ## Table of contents
