+++ docs/src/pages/components/Badge/AndroidBadge.mdx
@@ -5,12 +5,12 @@ documentationId: net.skyscanner.backpack.badge
 githubPath: badge
 ---
 
-import Readme from '../../../components/Readme';
-import AndroidComponentScreenshots from '../../../components/Screenshots/AndroidComponentScreenshots';
+import Readme from 'components/Readme';
+import AndroidComponentScreenshots from 'components/Screenshots/AndroidComponentScreenshots';
 
-import readme from '../../../../../backpack-android/docs/Badge/README.md';
-import screenshotAll from '../../../../../backpack-android/docs/Badge/screenshots/all.png';
-import screenshotAllDm from '../../../../../backpack-android/docs/Badge/screenshots/all_dm.png';
+import readme from 'backpack-android/docs/Badge/README.md';
+import screenshotAll from 'backpack-android/docs/Badge/screenshots/all.png';
+import screenshotAllDm from 'backpack-android/docs/Badge/screenshots/all_dm.png';
 
 ## Table of contents
