+++ docs/src/pages/components/Badge/IOSBadge.mdx
@@ -5,12 +5,12 @@ documentationId: BPKBadge
 githubPath: Badge
 ---
 
-import Readme from '../../../components/Readme';
-import IOSComponentScreenshots from '../../../components/Screenshots/IOSComponentScreenshots';
+import Readme from 'components/Readme';
+import IOSComponentScreenshots from 'components/Screenshots/IOSComponentScreenshots';
 
-import readme from '../../../../../backpack-ios/Backpack/Badge/README.md';
-import screenshotAll from '../../../../../backpack-ios/screenshots/iPhone 8-badge___all_lm.png';
-import screenshotAllDm from '../../../../../backpack-ios/screenshots/iPhone 8-badge___all_dm.png';
+import readme from 'backpack-ios/Backpack/Badge/README.md';
+import screenshotAll from 'backpack-ios/screenshots/iPhone 8-badge___all_lm.png';
+import screenshotAllDm from 'backpack-ios/screenshots/iPhone 8-badge___all_dm.png';
 
 ## Table of contents
