+++ docs/src/pages/components/Badge/NativeBadge.mdx
@@ -1,26 +1,26 @@
 ---
-title: Text
+title: Badge
 platform: native
 githubPath: bpk-component-badge
 ---
 
-import Readme from '../../../components/Readme';
-import NativeComponentScreenshots from '../../../components/Screenshots/NativeComponentScreenshots';
+import Readme from 'components/Readme';
+import NativeComponentScreenshots from 'components/Screenshots/NativeComponentScreenshots';
 
-import readme from '../../../../../backpack-react-native/lib/bpk-component-badge/README.md';
+import readme from 'backpack-react-native/lib/bpk-component-badge/README.md';
 
-import iosBadgeDefault from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/ios/default.png';
-import iosBadgeWithIcon from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/ios/with-icon.png';
-import iosBadgeWithTrailingIcon from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/ios/with-trailing-icon.png';
-import iosBadgeWithMultipleIcons from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/ios/with-multiple-icons.png';
-import iosBadgeDockedStart from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/ios/docked-start.png';
-import iosBadgeDockedEnd from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/ios/docked-end.png';
-import androidBadgeDefault from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/android/default.png';
-import androidBadgeWithIcon from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/android/with-icon.png';
-import androidBadgeWithTrailingIcon from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/android/with-trailing-icon.png';
-import androidBadgeWithMultipleIcons from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/android/with-multiple-icons.png';
-import androidBadgeDockedStart from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/android/docked-start.png';
-import androidBadgeDockedEnd from '../../../../../backpack-react-native/lib/bpk-component-badge/screenshots/android/docked-end.png';
+import iosBadgeDefault from 'backpack-react-native/lib/bpk-component-badge/screenshots/ios/default.png';
+import iosBadgeWithIcon from 'backpack-react-native/lib/bpk-component-badge/screenshots/ios/with-icon.png';
+import iosBadgeWithTrailingIcon from 'backpack-react-native/lib/bpk-component-badge/screenshots/ios/with-trailing-icon.png';
+import iosBadgeWithMultipleIcons from 'backpack-react-native/lib/bpk-component-badge/screenshots/ios/with-multiple-icons.png';
+import iosBadgeDockedStart from 'backpack-react-native/lib/bpk-component-badge/screenshots/ios/docked-start.png';
+import iosBadgeDockedEnd from 'backpack-react-native/lib/bpk-component-badge/screenshots/ios/docked-end.png';
+import androidBadgeDefault from 'backpack-react-native/lib/bpk-component-badge/screenshots/android/default.png';
+import androidBadgeWithIcon from 'backpack-react-native/lib/bpk-component-badge/screenshots/android/with-icon.png';
+import androidBadgeWithTrailingIcon from 'backpack-react-native/lib/bpk-component-badge/screenshots/android/with-trailing-icon.png';
+import androidBadgeWithMultipleIcons from 'backpack-react-native/lib/bpk-component-badge/screenshots/android/with-multiple-icons.png';
+import androidBadgeDockedStart from 'backpack-react-native/lib/bpk-component-badge/screenshots/android/docked-start.png';
+import androidBadgeDockedEnd from 'backpack-react-native/lib/bpk-component-badge/screenshots/android/docked-end.png';
 
 ## Table of contents
