+++ docs/src/pages/components/BannerAlert/NativeBannerAlert.mdx
@@ -4,16 +4,16 @@ platform: native
 githubPath: bpk-component-banner-alert
 ---
 
-import Readme from '../../../components/Readme';
-import NativeComponentScreenshots from '../../../components/Screenshots/NativeComponentScreenshots';
+import Readme from 'components/Readme';
+import NativeComponentScreenshots from 'components/Screenshots/NativeComponentScreenshots';
 
-import readme from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/README.md';
-import iosBannerAlertDefault from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/screenshots/ios/default.png';
-import iosBannerAlertDismissable from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/screenshots/ios/dismissable.png';
-import iosBannerAlertExpandable from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/screenshots/ios/expandable.png';
-import androidBannerAlertDefault from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/screenshots/android/default.png';
-import androidBannerAlertDismissable from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/screenshots/android/dismissable.png';
-import androidBannerAlertExpandable from '../../../../../backpack-react-native/lib/bpk-component-banner-alert/screenshots/android/expandable.png';
+import readme from 'backpack-react-native/lib/bpk-component-banner-alert/README.md';
+import iosBannerAlertDefault from 'backpack-react-native/lib/bpk-component-banner-alert/screenshots/ios/default.png';
+import iosBannerAlertDismissable from 'backpack-react-native/lib/bpk-component-banner-alert/screenshots/ios/dismissable.png';
+import iosBannerAlertExpandable from 'backpack-react-native/lib/bpk-component-banner-alert/screenshots/ios/expandable.png';
+import androidBannerAlertDefault from 'backpack-react-native/lib/bpk-component-banner-alert/screenshots/android/default.png';
+import androidBannerAlertDismissable from 'backpack-react-native/lib/bpk-component-banner-alert/screenshots/android/dismissable.png';
+import androidBannerAlertExpandable from 'backpack-react-native/lib/bpk-component-banner-alert/screenshots/android/expandable.png';
 
 
 ## Table of contents
