+++ docs/src/pages/components/Barcharts/AndroidBarChart.mdx
@@ -5,13 +5,13 @@ documentationId: net.skyscanner.backpack.barchart
 githubPath: barchart
 ---
 
-import Readme from '../../../components/Readme';
-import AndroidComponentScreenshots from '../../../components/Screenshots/AndroidComponentScreenshots';
+import Readme from 'components/Readme';
+import AndroidComponentScreenshots from 'components/Screenshots/AndroidComponentScreenshots';
 
-import readme from '../../../../../backpack-android/docs/BarChart/README.md';
+import readme from 'backpack-android/docs/BarChart/README.md';
 
-import screenshotDefault from '../../../../../backpack-android/docs/BarChart/screenshots/default.png';
-import screenshotDefaultDm from '../../../../../backpack-android/docs/BarChart/screenshots/default_dm.png';
+import screenshotDefault from 'backpack-android/docs/BarChart/screenshots/default.png';
+import screenshotDefaultDm from 'backpack-android/docs/BarChart/screenshots/default_dm.png';
 
 
 ## Table of contents
