+++ docs/src/pages/components/Barcharts/IOSBarChart.mdx
@@ -5,13 +5,13 @@ documentationId: BPKBarChart
 githubPath: BarChart
 ---
 
-import Readme from '../../../components/Readme';
-import IOSComponentScreenshots from '../../../components/Screenshots/IOSComponentScreenshots';
+import Readme from 'components/Readme';
+import IOSComponentScreenshots from 'components/Screenshots/IOSComponentScreenshots';
 
-import readme from '../../../../../backpack-ios/Backpack/BarChart/README.md';
+import readme from 'backpack-ios/Backpack/BarChart/README.md';
 
-import screenshotDefault from '../../../../../backpack-ios/screenshots/iPhone 8-bar-chart___default_lm.png';
-import screenshotDefaultDm from '../../../../../backpack-ios/screenshots/iPhone 8-bar-chart___default_dm.png';
+import screenshotDefault from 'backpack-ios/screenshots/iPhone 8-bar-chart___default_lm.png';
+import screenshotDefaultDm from 'backpack-ios/screenshots/iPhone 8-bar-chart___default_dm.png';
 
 ## Table of contents
