+++ docs/src/pages/components/BottomNav/AndroidBottomNav.mdx
@@ -5,13 +5,13 @@ documentationId: net.skyscanner.backpack.bottomnav
 githubPath: bottomnav
 ---
 
-import Readme from '../../../components/Readme';
-import AndroidComponentScreenshots from '../../../components/Screenshots/AndroidComponentScreenshots';
+import Readme from 'components/Readme';
+import AndroidComponentScreenshots from 'components/Screenshots/AndroidComponentScreenshots';
 
-import readme from '../../../../../backpack-android/docs/BottomNav/README.md';
+import readme from 'backpack-android/docs/BottomNav/README.md';
 
-import screenshotDefault from '../../../../../backpack-android/docs/BottomNav/screenshots/default.png';
-import screenshotDefaultDm from '../../../../../backpack-android/docs/BottomNav/screenshots/default_dm.png';
+import screenshotDefault from 'backpack-android/docs/BottomNav/screenshots/default.png';
+import screenshotDefaultDm from 'backpack-android/docs/BottomNav/screenshots/default_dm.png';
 
 ## Table of contents
