+++ docs/src/pages/components/BottomSheet/BottomSheetPage.js
@@ -20,11 +20,11 @@
 
 import React from 'react';
 
-import DocsPageWrapper from '../../../components/DocsPageWrapper';
-import IntroBlurb from '../../../components/IntroBlurb';
-import MarkdownPage from '../../../components/MarkdownPage';
-
 import IOS, { metadata as iosMetadata } from './IOSBottomSheet.mdx';
+
+import DocsPageWrapper from 'components/DocsPageWrapper';
+import IntroBlurb from 'components/IntroBlurb';
+import MarkdownPage from 'components/MarkdownPage';
 
 const BottomSheetPage = () => (
   <DocsPageWrapper
