+++ docs/src/pages/components/BottomSheet/IOSBottomSheet.mdx
@@ -5,13 +5,13 @@ documentationId: BPKBottomSheet
 githubPath: BottomSheet
 ---
 
-import Readme from '../../../components/Readme';
-import IOSComponentScreenshots from '../../../components/Screenshots/IOSComponentScreenshots';
+import Readme from 'components/Readme';
+import IOSComponentScreenshots from 'components/Screenshots/IOSComponentScreenshots';
 
-import readme from '../../../../../backpack-ios/Backpack/BottomSheet/README.md';
+import readme from 'backpack-ios/Backpack/BottomSheet/README.md';
 
-import screenshotAll from '../../../../../backpack-ios/screenshots/iPhone 8-bottom-sheet___with-bottom-section_lm.png';
-import screenshotAllDm from '../../../../../backpack-ios/screenshots/iPhone 8-bottom-sheet___with-bottom-section_dm.png';
+import screenshotAll from 'backpack-ios/screenshots/iPhone 8-bottom-sheet___with-bottom-section_lm.png';
+import screenshotAllDm from 'backpack-ios/screenshots/iPhone 8-bottom-sheet___with-bottom-section_dm.png';
 
 ## Table of contents
