+++ docs/src/pages/components/Button/AndroidButton.mdx
@@ -5,24 +5,24 @@ documentationId: net.skyscanner.backpack.button
 githubPath: button
 ---
 
-import Readme from '../../../components/Readme';
-import AndroidComponentScreenshots from '../../../components/Screenshots/AndroidComponentScreenshots';
+import Readme from 'components/Readme';
+import AndroidComponentScreenshots from 'components/Screenshots/AndroidComponentScreenshots';
 
-import buttonReadme from '../../../../../backpack-android/docs/Button/README.md';
-import linkReadme from '../../../../../backpack-android/docs/ButtonLink/README.md';
+import buttonReadme from 'backpack-android/docs/Button/README.md';
+import linkReadme from 'backpack-android/docs/ButtonLink/README.md';
 
-import screenshotPrimary from '../../../../../backpack-android/docs/Button/screenshots/primary.png';
-import screenshotPrimaryDm from '../../../../../backpack-android/docs/Button/screenshots/primary_dm.png';
-import screenshotSecondary from '../../../../../backpack-android/docs/Button/screenshots/secondary.png';
-import screenshotSecondaryDm from '../../../../../backpack-android/docs/Button/screenshots/secondary_dm.png';
-import screenshotDestructive from '../../../../../backpack-android/docs/Button/screenshots/destructive.png';
-import screenshotDestructiveDm from '../../../../../backpack-android/docs/Button/screenshots/destructive_dm.png';
-import screenshotOutline from '../../../../../backpack-android/docs/Button/screenshots/outline.png';
-import screenshotOutlineDm from '../../../../../backpack-android/docs/Button/screenshots/outline_dm.png';
-import screenshotFeatured from '../../../../../backpack-android/docs/Button/screenshots/featured.png';
-import screenshotFeaturedDm from '../../../../../backpack-android/docs/Button/screenshots/featured_dm.png';
-import screenshotLink from '../../../../../backpack-android/docs/ButtonLink/screenshots/default.png';
-import screenshotLinkDm from '../../../../../backpack-android/docs/ButtonLink/screenshots/default_dm.png';
+import screenshotPrimary from 'backpack-android/docs/Button/screenshots/primary.png';
+import screenshotPrimaryDm from 'backpack-android/docs/Button/screenshots/primary_dm.png';
+import screenshotSecondary from 'backpack-android/docs/Button/screenshots/secondary.png';
+import screenshotSecondaryDm from 'backpack-android/docs/Button/screenshots/secondary_dm.png';
+import screenshotDestructive from 'backpack-android/docs/Button/screenshots/destructive.png';
+import screenshotDestructiveDm from 'backpack-android/docs/Button/screenshots/destructive_dm.png';
+import screenshotOutline from 'backpack-android/docs/Button/screenshots/outline.png';
+import screenshotOutlineDm from 'backpack-android/docs/Button/screenshots/outline_dm.png';
+import screenshotFeatured from 'backpack-android/docs/Button/screenshots/featured.png';
+import screenshotFeaturedDm from 'backpack-android/docs/Button/screenshots/featured_dm.png';
+import screenshotLink from 'backpack-android/docs/ButtonLink/screenshots/default.png';
+import screenshotLinkDm from 'backpack-android/docs/ButtonLink/screenshots/default_dm.png';
 
 ## Table of contents
