+++ docs/src/pages/components/Button/ButtonPage.js
@@ -20,14 +20,14 @@
 
 import React from 'react';
 
-import DocsPageWrapper from '../../../components/DocsPageWrapper';
-import IntroBlurb from '../../../components/IntroBlurb';
-import MarkdownPage from '../../../components/MarkdownPage';
-
 import Android, { metadata as androidMetadata } from './AndroidButton.mdx';
 import IOS, { metadata as iosMetadata } from './IOSButtonPage.mdx';
 import WebButtons from './WebButtonPage';
 import Native, { metadata as nativeMetadata } from './NativeButtonPage.mdx';
+
+import MarkdownPage from 'components/MarkdownPage';
+import IntroBlurb from 'components/IntroBlurb';
+import DocsPageWrapper from 'components/DocsPageWrapper';
 
 const ButtonPage = () => (
   <DocsPageWrapper
