+++ docs/src/pages/components/Button/NativeButtonPage.mdx
@@ -4,21 +4,21 @@ platform: native
 githubPath: bpk-component-text
 ---
 
-import Readme from '../../../components/Readme';
-import NativeComponentScreenshots from '../../../components/Screenshots/NativeComponentScreenshots';
+import Readme from 'components/Readme';
+import NativeComponentScreenshots from 'components/Screenshots/NativeComponentScreenshots';
 
-import buttonReadme from '../../../../../backpack-react-native/lib/bpk-component-button/README.md';
-import linkReadme from '../../../../../backpack-react-native/lib/bpk-component-button-link/README.md';
-import iosButtonPrimary from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/ios/primary.png';
-import iosButtonSecondary from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/ios/secondary.png';
-import iosButtonDestructive from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/ios/destructive.png';
-import iosButtonFeatured from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/ios/featured.png';
-import iosButtonLink from '../../../../../backpack-react-native/lib/bpk-component-button-link/screenshots/ios/default.png';
-import androidButtonPrimary from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/android/primary.png';
-import androidButtonSecondary from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/android/secondary.png';
-import androidButtonDestructive from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/android/destructive.png';
-import androidButtonFeatured from '../../../../../backpack-react-native/lib/bpk-component-button/screenshots/android/featured.png';
-import androidButtonLink from '../../../../../backpack-react-native/lib/bpk-component-button-link/screenshots/android/default.png';
+import buttonReadme from 'backpack-react-native/lib/bpk-component-button/README.md';
+import linkReadme from 'backpack-react-native/lib/bpk-component-button-link/README.md';
+import iosButtonPrimary from 'backpack-react-native/lib/bpk-component-button/screenshots/ios/primary.png';
+import iosButtonSecondary from 'backpack-react-native/lib/bpk-component-button/screenshots/ios/secondary.png';
+import iosButtonDestructive from 'backpack-react-native/lib/bpk-component-button/screenshots/ios/destructive.png';
+import iosButtonFeatured from 'backpack-react-native/lib/bpk-component-button/screenshots/ios/featured.png';
+import iosButtonLink from 'backpack-react-native/lib/bpk-component-button-link/screenshots/ios/default.png';
+import androidButtonPrimary from 'backpack-react-native/lib/bpk-component-button/screenshots/android/primary.png';
+import androidButtonSecondary from 'backpack-react-native/lib/bpk-component-button/screenshots/android/secondary.png';
+import androidButtonDestructive from 'backpack-react-native/lib/bpk-component-button/screenshots/android/destructive.png';
+import androidButtonFeatured from 'backpack-react-native/lib/bpk-component-button/screenshots/android/featured.png';
+import androidButtonLink from 'backpack-react-native/lib/bpk-component-button-link/screenshots/android/default.png';
 
 ## Table of contents
 
@@ -53,14 +53,6 @@ import androidButtonLink from '../../../../../backpack-react-native/lib/bpk-comp
   iOSAltText="iOS featured button"
   android={androidButtonFeatured}
   androidAltText="Android featured button" />
-
-## Outline
-
-<NativeComponentScreenshots
-  iOS={iosButtonOutline}
-  iOSAltText="iOS outline button"
-  android={androidButtonOutline}
-  androidAltText="Android outline button" />
 
 ## Button Implementation
