+++ docs/src/pages/components/Button/WebButtonPage.js
@@ -38,11 +38,12 @@ import loadingButtonReadme from 'bpk-component-loading-button/README.md';
 import { cssModules } from 'bpk-react-utils';
 
 import * as ROUTES from '../../../constants/routes';
-import { WebComponentPage } from '../../../components/ComponentPage';
-import Paragraph from '../../../components/Paragraph';
-import Code from '../../../components/Code';
 
 import STYLES from './buttons-page.scss';
+
+import { WebComponentPage } from 'components/ComponentPage';
+import Paragraph from 'components/Paragraph';
+import Code from 'components/Code';
 
 const AlignedBpkSmallArrowIcon = alignToButton(TestBpkSmallArrowIcon);
 const AlignedBpkLargeArrowIcon = alignToLargeButton(TestBpkLargeArrowIcon);
